backend/requirements.txt
# ── Core dependencies ──
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.9
httpx==0.27.0
google-generativeai==0.7.2
# Resume parsing
pdfplumber==0.11.0
python-docx==1.1.0
# Scraping
beautifulsoup4==4.12.3
lxml==5.2.1
numpy==1.26.4
python-dotenv==1.0.1
email-validator==2.1.1
scikit-learn>=1.4.0

# ── Optional: Advanced NLP (install separately if needed) ──
# These have a dependency conflict with FastAPI (typer version).
# The app works without them — all NLP features have graceful fallbacks.
# To install: pip install spacy transformers sentence-transformers torch
# spacy==3.7.4
# transformers==4.40.0
# sentence-transformers==2.7.0
# torch>=2.2.0